integrating updated version of aggsig, which gives greater control over the contents of e
